feat: Add api error model for RFC9457 (#206)
* feat: implement standardized error response model and update API documentation

* feat: enhance error handling in agent listing and update API documentation

* feat: standardize error response URIs in agent and agent group controllers

* wip

* test

* chore: fix lint

* remove test

* fix codecov

* fix typo

+++ internal/domain/service/connection.go
@@ -116,7 +116,7 @@ func (s *Service) ListConnections(
 	totalMatchedItemsCount := len(keys)
 
 	var nextContinue string
-	if options.Limit > 0 && len(keys) > int(options.Limit) {
+	if options.Limit > 0 && int64(len(keys)) > options.Limit {
 		nextContinue = keys[options.Limit]
 		keys = keys[:options.Limit]
 	} else {
